requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
coverage==6.4.2
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.50.0
django-extensions==3.2.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
drf-spectacular==0.23.1
executing==0.9.1
factory-boy==3.2.1
Faker==13.15.1
filelock==3.7.1
flake8==5.0.4
identify==2.5.3
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
jsonschema==4.9.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
types-pytz==2022.1.2
types-PyYAML==6.0.11
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.3
wcwidth==0.2.5
zipp==3.8.1
